Move content of interview_portal modules to dashboard modules
