+++ dashboard/admin.py
@@ -1,3 +1,20 @@
 from django.contrib import admin
+from .models import (
+    Technology,
+    Status,
+    Application,
+    ApplicationDetail
+)
 
-# Register your models here.
+
+class ApplicationDetailInline(admin.StackedInline):
+    model = ApplicationDetail
+
+
+@admin.register(Application)
+class ApplicationAdmin(admin.ModelAdmin):
+    inlines = [ApplicationDetailInline, ]
+
+
+admin.site.register(Technology)
+admin.site.register(Status)
