+++ interview_portal/urls.py
@@ -17,16 +17,13 @@ from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
-from django.views.generic import TemplateView
-from interview_portal import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
 
-    # Authentication
+    # Allauth URLs
     path('accounts/', include('allauth.urls')),
-    path('login/', views.CustomLoginView.as_view(), name='login'),
 
     # Dashboard
-    path('', TemplateView.as_view(template_name='interview_portal/index.html'), name='home'),
+    path('/', include('dashboard.urls', namespace='dashboard')),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
